refactor: extract sanitizeOptionLabel to lib/utils for testability
Move the pure sanitization function out of assistant-architect-streaming.tsx
(which imports @assistant-ui/react and other heavy UI deps) into a standalone
lib/utils/sanitize-option-label.ts that Jest can import without resolution
failures in the jsdom test environment.

The streaming file re-exports the function to preserve its public API;
the test file's import is updated to point to the utility module directly.
